Add identity razor pages and move react router to index
