+++ WhaleSpotting/WhaleSpotting.csproj
@@ -11,16 +11,12 @@
 
     <ItemGroup>
         <PackageReference Include="Microsoft.AspNetCore.ApiAuthorization.IdentityServer" Version="5.0.8" />
+        <PackageReference Include="Microsoft.AspNetCore.Identity.EntityFrameworkCore" Version="5.0.8" />
+        <PackageReference Include="Microsoft.AspNetCore.Identity.UI" Version="5.0.8" />
         <PackageReference Include="Microsoft.AspNetCore.SpaServices.Extensions" Version="5.0.8" />
         <PackageReference Include="Microsoft.EntityFrameworkCore" Version="5.0.8" />
-        <PackageReference Include="Microsoft.EntityFrameworkCore.Design" Version="5.0.8">
-          <PrivateAssets>all</PrivateAssets>
-          <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
-        </PackageReference>
-        <PackageReference Include="Microsoft.EntityFrameworkCore.Tools" Version="5.0.8">
-          <PrivateAssets>all</PrivateAssets>
-          <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
-        </PackageReference>
+        <PackageReference Include="Microsoft.EntityFrameworkCore.Design" Version="5.0.8" />
+        <PackageReference Include="Microsoft.EntityFrameworkCore.Tools" Version="5.0.8" />
         <PackageReference Include="Microsoft.EntityFrameworkCore.Relational" Version="5.0.8" />
         <PackageReference Include="Npgsql.EntityFrameworkCore.PostgreSQL" Version="5.0.7" />
     </ItemGroup>
