+++ WhaleSpotting/appsettings.json
@@ -8,10 +8,8 @@
   },
   "IdentityServer": {
     "Clients": {
-      "react_app": {
-        "Profile": "SPA",
-        "RedirectUri": "https://www.example.com/authentication/login-callback",
-        "LogoutUri": "https://www.example.com/authentication/logout-callback"
+      "WhaleSpotting": {
+        "Profile": "IdentityServerSPA"
       }
     }
   },
