+++ WhaleSpotting/react-app/src/index.tsx
@@ -5,11 +5,14 @@ import App from "./App";
 import reportWebVitals from "./reportWebVitals";
 import registerServiceWorker from "./registerServiceWorker";
 import "bootstrap/dist/css/bootstrap.css";
+import { BrowserRouter } from "react-router-dom";
+
+const baseUrl = document.getElementsByTagName("base")[0].getAttribute("href") as string;
 
 ReactDOM.render(
-    <React.StrictMode>
+    <BrowserRouter basename={baseUrl}>
         <App />
-    </React.StrictMode>,
+    </BrowserRouter>,
     document.getElementById("root")
 );
